+++ tests/test_stats/import_vector/import_vector_temporal.json
@@ -2,8 +2,8 @@
   "eopatch-id-0-col-0-row-0": {
     "bbox": "BBox(((729480.0, 4390045.0), (732120.0, 4391255.0)), crs=CRS('32638'))",
     "timestamps": [
-      "2016-11-05T00:00:00",
-      "2016-11-01T00:00:00"
+      "2016-11-01T00:00:00",
+      "2016-11-05T00:00:00"
     ],
     "vector": {
       "crop": {
